+++ mvc/src/main/java/com/interface21/webmvc/servlet/HandlerMappingRegistry.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet;
 
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
