+++ mvc/src/test/java/com/interface21/webmvc/servlet/HandlerMappingRegistryTest.java
@@ -1,9 +1,12 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+import com.interface21.webmvc.servlet.HandlerMapping;
+import com.interface21.webmvc.servlet.handler.HandlerExecution;
+import com.interface21.webmvc.servlet.HandlerMappingRegistry;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
